fix many warnings, improve tests
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/6799


git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@25772 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/i18n/calendar.cpp
@@ -262,7 +262,7 @@ static ECalType getCalendarTypeForLocale(const char *locid) {
         // the first calender type is the default for the region
         int32_t len = 0;
         const UChar *uCalType = ures_getStringByIndex(order, 0, &len, &status);
-        if (len < sizeof(calTypeBuf)) {
+        if (len < (int32_t)sizeof(calTypeBuf)) {
             u_UCharsToChars(uCalType, calTypeBuf, len);
             *(calTypeBuf + len) = 0; // terminate;
             calType = getCalendarType(calTypeBuf);
